requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.1.3
certifi==2024.7.4
click==8.1.7
dnspython==2.6.1
ecdsa==0.19.0
email-validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
jinja2==3.1.4
mako==1.3.5
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
orjson==3.10.6
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.0
pydantic==2.8.2
pydantic-core==2.20.1
pygments==2.18.0
pyjwt==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pyyaml==6.0.1
rich==13.7.1
rsa==4.9
setuptools==70.3.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.29
sqlalchemy-serializer==1.4.22
starlette==0.37.2
typer==0.12.3
typing-extensions==4.12.2
ujson==5.10.0
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
